+++ tests/Controller/Support/AvdlControllerTest.php
@@ -95,7 +95,7 @@ class AvdlControllerTest extends WebTestCase
         $this->createLogin($user);
 
         $id = $this->fixtures['peopleGroup3']->getId();
-        $this->client->request('POST', "/group/$id/support/new", [
+        $this->client->request('POST', "/people-group/$id/new-support", [
             'support' => [
                 'service' => $this->fixtures['serviceAvdl'],
                 'device' => $this->fixtures['deviceAvdl']->getCode(),